+++ components/date-picker/BasePicker.js
@@ -273,7 +273,7 @@ class BasePicker extends Component {
     }
   }
   _input (text, ref, placeholder) {
-    const {disabled} = this.props
+    const {disabled, inputReadOnly} = this.props
     const { texts } = this.state
     return (
       <input
@@ -282,6 +282,7 @@ class BasePicker extends Component {
         placeholder={placeholder}
         className={disabled ? 'disabled' : ''}
         disabled={disabled}
+        readonly={inputReadOnly ? 'readOnly' : inputReadOnly}
         onChange={e => {
           ref === 'input' ? (texts[0] = e.target.value) : (texts[1] = e.target.value)
           this.setState({