+++ components/date-picker/Time.js
@@ -78,11 +78,34 @@ class Time extends Component {
 
     return isDisabled
   }
+  
+  getStep(type){
+    const {hourStep,minuteStep,secondStep} = this.props
+    let step = 1
+    const directionStep = 1
+    switch (type) {
+      case 'hours':
+        step = hourStep || directionStep
+        break;
+      case 'minutes':
+        step = minuteStep || directionStep
+        break;
+      case 'seconds':
+        step = secondStep || directionStep
+        break;
+      default:
+        step = 1
+        break;
+    }
+    return step
+  }
+
   generateDatas (type) {
 
     let count
     let datas = []
     const currentDate = deconstructDate(this.state.date)
+    const step = this.getStep(type)
     const disabledList = this._getDsiabledList()
     const disabledTime = disabledList[type]
     if (type === 'hours') {
@@ -93,12 +116,14 @@ class Time extends Component {
       count = 60
     }
     for (let i = 0; i < count; i += 1) {
-      datas.push({
-        value: i,
-        text: i < 10 ? '0' + i : i.toString(),
-        disabled: this.setDisableTime(type,i,disabledTime),
-        current: i === currentDate[type]
-      })
+      if(i%step === 0 || i === 0){
+        datas.push({
+          value: i,
+          text: i < 10 ? '0' + i : i.toString(),
+          disabled: this.setDisableTime(type,i,disabledTime),
+          current: i === currentDate[type]
+        })
+      }
     }
     return datas
   }