+++ components/date-picker/TimeList.js
@@ -23,8 +23,9 @@ export default class TimeList extends Component {
     this.topValue_2 = 0
   }
   scrollTo () {
+    const arrow = this.getStep()
     const { value } = this.props
-    const dVal = 32
+    const dVal = 32/arrow
     this.listRef.current && (this.listRef.current.scrollTop = value * dVal)
   }
   componentDidUpdate () {
@@ -36,22 +37,46 @@ export default class TimeList extends Component {
       this.scrollTo()
     }, 0)
   }
+
   componentWillUnmount () {
     window.clearTimeout(this.timer)
     // this.listRef.current.removeEventListener('scroll', this.scrollEvent)
+  }
+
+  getStep(direction){
+    const {hourStep,minuteStep,secondStep,type} = this.props
+    const directionStep = direction && direction === 'up' ? -1 : 1
+    let step = directionStep
+
+    switch (type) {
+      case 'hours':
+        step = hourStep || directionStep
+        break;
+      case 'minutes':
+        step = minuteStep || directionStep
+        break;
+      case 'seconds':
+        step = secondStep || directionStep
+        break;
+      default:
+        step = directionStep
+        break;
+    }
+
+    return step
   }
   renderArrow (type) {
     return (
       <React.Fragment>
         <span
           className='hi-timepicker__page-turn'
-          onClick={() => this.arrowEvent(-1)}
+          onClick={() => this.arrowEvent('up')}
         >
           <Icon name='up' />
         </span>
         <span
           className='hi-timepicker__page-turn'
-          onClick={() => this.arrowEvent(1)}
+          onClick={() => this.arrowEvent('down')}
         >
           <Icon name='down' />
         </span>
@@ -68,19 +93,21 @@ export default class TimeList extends Component {
     })
     return isDisabled
   }
-  arrowEvent (arrow) {
+  arrowEvent (direction) {
+    const arrow = this.getStep(direction)
     const st = this.listRef.current.scrollTop
-    const val = Math.round(st / 32)
+    const val = Math.round(st * arrow / 32)
     !this.getdisabledType(val) && this.props.onSelect(this.props.type, val + arrow, arrow)
   }
 
   isScrollStop (val, el) {
     const { disabledList } = this.props
-    this.topValue_2 = el.scrollTop
+    this.topValue_2 = el.scrollTop 
     if (this.topValue_1 === this.topValue_2) {
       el.scrollTop = val * 32
       if (!disabledList.includes(val)) {
-        !this.getdisabledType(val) && this.props.onSelect(this.props.type, val)
+        const arrow = this.getStep()
+        !this.getdisabledType(val)  && this.props.onSelect(this.props.type, val * arrow)
       }
     }
   }
@@ -95,6 +122,7 @@ export default class TimeList extends Component {
     }
     this.timer = setTimeout(this.isScrollStop.bind(this, val, e.target), 200)
   }
+
   clickEvent (type, e) {
     e.stopPropagation()
     const li = e.target